Allowed the driver to control clock token delivery
The `ClockBridge` now exposes two registers, `CREDIT` and `CREDIT_COUNT` which
can be used to set and query the number of clock tokens which the bridge is
allowed to deliver to the target. Tests which rely on `PeekPokeBridge` credit
the system with a maximal number of tokens. `firesim_top` also infinitely credits
the system alongside a dummy `PeekPokeBridge`, but in future PRs the sole control
mechanism should be that through the clock bridge.

+++ sim/firesim-lib/src/main/scala/bridges/SerialBridge.scala
@@ -100,7 +100,7 @@ class SerialBridgeModule(serialBridgeParams: SerialBridgeParams)(implicit p: Par
           "serial",
           Seq(
               CppBoolean(serialBridgeParams.memoryRegionNameOpt.isDefined),
-              UInt32(offsetConst)
+              Int64(offsetConst)
           ),
           hasLoadMem = true
       )
